color refactor, drawLine method

+++ jena/engine/graphics/ColorFloatStruct.java
@@ -4,9 +4,9 @@ public class ColorFloatStruct extends ColorStruct
 {
     public ColorFloatStruct(float r, float g, float b, float a)
     {
-        this.r = (byte)(r * float2byte);
-        this.g = (byte)(g * float2byte);
-        this.b = (byte)(b * float2byte);
-        this.a = (byte)(a * float2byte);
+        this.r = (int)(r * float2byte);
+        this.g = (int)(g * float2byte);
+        this.b = (int)(b * float2byte);
+        this.a = (int)(a * float2byte);
     }
 }
